Encrypt my secrets

+++ index.js
@@ -2,18 +2,24 @@
 
 'use strict'
 
+require('dotenv').config()
+
 const minimist = require('minimist')
+const promptly = require('promptly')
 const argv = minimist(process.argv.slice(2))
 const { createDB } = require('./lib')
 
+const prompPassword = () => promptly.password('Enter your password: ', { replace: '*' })
+
 async function main () {
-  const db = await createDB('postgres')
+  const db = await createDB(process.env.DB_TYPE)
   const command = argv._.shift()
 
   switch (command) {
     case 'users:create':
       try {
-        const { user, pass } = argv
+        const { user } = argv
+        const pass = await prompPassword()
 
         await db.createUser(user, pass)
         console.log(`User ${user} created!`)
@@ -36,16 +42,24 @@ async function main () {
     case 'secrets:create':
       try {
         const { user, name, value } = argv
-        await db.createSecret(user, name, value)
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+
+        await db.createSecret(user, pass, name, value)
 
         console.log(`Secret ${name} created`)
       } catch (err) {
+        console.log(err)
         throw new Error('Cannot create secret')
       }
       break
     case 'secrets:list':
       try {
         const { user } = argv
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
         const secrets = await db.listSecrets(user)
 
         secrets.forEach(secret => {
@@ -58,7 +72,10 @@ async function main () {
     case 'secrets:get':
       try {
         const { user, name } = argv
-        const secret = await db.getSecret(user, name)
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+        const secret = await db.getSecret(user, pass, name)
 
         if (!secret) return console.log(`Secret ${name} not found`)
         console.log(`- ${secret.name} = ${secret.value}`)
@@ -69,6 +86,9 @@ async function main () {
     case 'secrets:update':
       try {
         const { user, name, value } = argv
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
         await db.updateSecret(user, name, value)
 
         console.log(`Secret ${name} updated`)
@@ -79,6 +99,9 @@ async function main () {
     case 'secrets:delete':
       try {
         const { user, name } = argv
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
         await db.deleteSecret(user, name)
 
         console.log(`Secret ${name} deleted`)
